feat(utils/evaluation.py): update sche metric hook for valid
