+++ internlm/core/scheduler/__init__.py
@@ -1,4 +1,4 @@
-from .base_scheduler import BaseScheduler, SchedulerHook
+from .base_scheduler import BaseScheduler, SchedulerHook, SchedulerMetricHook
 from .no_pipeline_scheduler import NonPipelineScheduler
 from .pipeline_scheduler import InterleavedPipelineScheduler, PipelineScheduler
 
@@ -8,4 +8,5 @@ __all__ = [
     "InterleavedPipelineScheduler",
     "PipelineScheduler",
     "SchedulerHook",
+    "SchedulerMetricHook",
 ]
