+++ internlm/core/scheduler/no_pipeline_scheduler.py
@@ -88,7 +88,6 @@ class NonPipelineScheduler(BaseScheduler):
         forward_only: bool = False,
         return_loss: bool = True,
         scale_loss: int = 1,
-        post_fn: Callable = None,
     ):
         """Trains one batch of data.
 
@@ -100,7 +99,6 @@ class NonPipelineScheduler(BaseScheduler):
                 be executed.
             return_loss (bool, optional): Loss will be returned if True.
             scale_loss (int, optional): The scale factor for the loss.
-            post_fn (Callable, optional): Call back function after executing data forward output.
         """
 
         # forward
@@ -110,9 +108,6 @@ class NonPipelineScheduler(BaseScheduler):
             self._call_hooks("after_forward", output)
 
             self._call_hooks("post_helper_func", output, label)
-
-            if post_fn is not None:
-                post_fn(output, label)
 
             if return_loss:
                 self._call_hooks("before_criterion", output, label)
@@ -138,7 +133,6 @@ class NonPipelineScheduler(BaseScheduler):
         forward_only: bool = False,
         return_loss: bool = True,
         return_output_label: bool = True,
-        post_fn: Callable = None,
     ):
         """The process function that loads a batch of dataset and feeds it to the model.
         The returned labels and loss will None if :attr:`return_loss` is False.
@@ -150,7 +144,6 @@ class NonPipelineScheduler(BaseScheduler):
                 If True, the model is run for the forward pass, else back propagation will be executed.
             return_loss (bool, optional): Loss will be returned if True.
             return_output_label (bool, optional): Output and label will be returned if True.
-            post_fn (Callable, optional): Call back function after executing data forward output.
 
         Returns:
             Tuple[:class:`torch.Tensor`]: A tuple of (output, label, loss), loss and label could be None.
@@ -184,7 +177,7 @@ class NonPipelineScheduler(BaseScheduler):
             _data, _label = self._load_accum_batch(data, label)
 
             _output, _loss = self._train_one_batch(
-                _data, _label, engine, forward_only, return_loss, self._grad_accum_size, post_fn
+                _data, _label, engine, forward_only, return_loss, self._grad_accum_size
             )
 
             if return_loss:
